apps/ai_models/base.py
# apps/ai_models/base.py
from abc import ABC, abstractmethod

class BaseDetector(ABC):
    """Interface pour tous les détecteurs"""
    
    @abstractmethod
    def analyze(self, file_path):
        pass

# apps/ai_models/detectors/deepfake_detector.py
import cv2
import numpy as np
from ..base import BaseDetector

class DeepfakeDetector(BaseDetector):
    def __init__(self):
        self.model = self.load_model()
    
    def load_model(self):
        """Charger le modèle préentraîné"""
        # Charger depuis TensorFlow/PyTorch
        pass
    
    def analyze(self, file_path):
        """Analyser une image"""
        image = cv2.imread(file_path)
        prediction = self.model.predict(image)
        
        return {
            'is_deepfake': prediction > 0.5,
            'confidence': float(prediction),
            'faces_detected': 1,
        }

# apps/ai_models/factories.py
class AIDetectorFactory:
    @staticmethod
    def create(media_type):
        detectors = {
            'image': DeepfakeDetector,
            'video': VideoAnalyzer,
            'audio': AudioAnalyzer,
        }
        return detectors[media_type]()
