requirements.txt
# Django & DRF
Django==4.2.0
djangorestframework==3.14.0
django-filter==23.1
django-cors-headers==4.0.0

# Database
psycopg2-binary==2.9.6
mysqlclient==2.1.1

# AI/ML
torch==2.0.0
torchvision==0.15.0
torchaudio==2.0.0
numpy==1.24.3
opencv-python==4.7.0.72
scikit-learn==1.2.2
scikit-image==0.20.0
Pillow==9.5.0

# Processing
imageio==2.25.1
imageio-ffmpeg==0.4.8
librosa==0.10.0
soundfile==0.12.1

# Utils
tqdm==4.65.0
requests==2.31.0
pandas==2.0.2
matplotlib==3.7.1
Werkzeug==2.3.0

# API & Serialization
pyyaml==6.0
jsonschema==4.17.3

# Development
pytest==7.3.1
pytest-django==4.5.2
black==23.3.0
flake8==6.0.0

# Production
gunicorn==20.1.0
celery==5.2.7
redis==4.5.5
python-dotenv==1.0.0